Hand dAqk and dAkk to the intra backward in the activation dtype
Both gradients are written once and consumed only by
chunk_precond_kda_bwd_kernel_intra, which casts every load to fp32 before
using it, so the fp32 buffers bought nothing: they doubled two 10.5 MB
writes and the roughly 5x amplified reads the intra kernel makes of each
(NC+1 = 5 [16,16] blocks per program, 16 programs per chunk). This is the
same trade the dq/dk/dkg hand-off already takes.

chunk_precond_kda_bwd_kernel_intra 0.2686 -> 0.2593 ms, wy_dqkg
0.2623 -> 0.2519 ms, dAv 0.0558 -> 0.0489 ms (mean of 10 iterations,
RTX 4090). PKDA layer fwd+bwd kernel total 2.3279 -> 2.2849 ms; max
relative gradient drift 7.438e-3 -> 7.479e-3.

+++ fla/ops/precond_kda/chunk_bwd.py
@@ -136,7 +136,10 @@ def chunk_precond_kda_bwd_dAv(
     BV = min(max(triton.next_power_of_2(V), 16), CONST_TILING)
     NT = triton.cdiv(T, BT) if cu_seqlens is None else len(chunk_indices)
 
-    dA = v.new_empty(B, T, H, BT, dtype=torch.float)
+    # dA is written once and read back only by the intra backward, which casts
+    # every load to fp32 before it is used; the activation dtype halves that
+    # buffer and the ~5x amplified reads the intra kernel makes of it.
+    dA = v.new_empty(B, T, H, BT)
     dv = torch.empty_like(do)
     grid = (NT, B * H)
     chunk_precond_kda_bwd_kernel_dAv[grid](
@@ -474,7 +477,7 @@ def chunk_precond_kda_bwd_wy_dqkg(
     dv2 = torch.empty_like(v)
     dg = torch.empty_like(g, dtype=torch.float)
     db = torch.empty_like(beta, dtype=torch.float)
-    dA = torch.empty_like(A, dtype=torch.float)
+    dA = torch.empty_like(A)
 
     grid = (NT, B * H)
     chunk_precond_kda_bwd_kernel_wy_dqkg[grid](
